+++ src/pages/Mypage/WatermarkPanel.tsx
@@ -1,4 +1,4 @@
-import RecordPage from "./RecordPage";
+import RecordPage from "../../components/Mypage/RecordPage";
 import { useNavigate } from "react-router-dom";
 
 const watermarkRecords = [
