+++ src/pages/Mypage/Withdraw.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import PasswordInput from "./PasswordInput";
+import PasswordInput from "../../components/Mypage/PasswordInput";
 
 function Withdraw() {
   const [password, setPassword] = useState("");
